SchedulerConfig moved to config package
